Update 29th April 12:11am
Completed the Design (although I am struggling to make the spacing look like the wireframe and have failed to add in the form into the background. Also a bit confused as to how to change the fonts within the text fields - will look into this later)

Still need to integrate firebase into the registration as well as completing a form validation for correct email and password confirmation.