Order confirmation
Just wanted to make a PR for this when I started doing staff list with this it started getting a bit confusing so thought I'd just do a PR now. So this is basically just a confirmation for all the order details. Once the reservation details are merged with order detials I'll finish it off.